remove workerize, use worker-loader+promiser-worker
