+++ routes/_database/database.js
@@ -1,3 +1,3 @@
 // dev version to play nice with IDEs
-import * as database from './databaseWorker'
+import * as database from './databaseApi.js'
 export { database }
