--- routes/_database/database.workerize.js
@@ -1,3 +1,0 @@
-// workerize version, used in production
-const database = process.browser && require('./databaseWorker.js')()
-export { database }
